update node settings.
