+++ src/App.tsx
@@ -19,7 +19,6 @@ import ReceiveMoney from 'Routes/Receive/ReceiveMoney'
 import { DataSyncProvider } from './providers/DataSyncProvider'
 import ElectronThemeChangeHandler from 'Components/ElectronThemeChangeHandler'
 import Initializing from 'Routes/Initializing'
-import NodeSettings from 'Routes/Settings/NodeSettings'
 
 const breakpoints = {
   xs: '46.875rem', //750px
@@ -58,7 +57,6 @@ function App() {
                 />
                 {/* <Route path={ROUTES.RESOURCES} element={null} /> */}
                 <Route path={ROUTES.NODE} element={<NodeOverview />} />
-                <Route path={ROUTES.NODE_SETTINGS} element={<NodeSettings />} />
                 {/* <Route path={ROUTES.MINER} element={<Miner />} /> */}
               </Route>
             </Route>
