+++ src/data/DemoNodeManager.ts
@@ -59,7 +59,7 @@ const PEERS: PeerResponse[] = Array(23)
   })
 
 let DEMO_NODE_CONFIG: Partial<ConfigOptions> = {
-  nodeName: '',
+  nodeName: STATUS.node.nodeName,
   blockGraffiti: '',
   nodeWorkers: 6,
   minPeers: -1,
@@ -145,6 +145,9 @@ class DemoNodeManager {
   async save(): Promise<void> {
     return new Promise(resolve => {
       setTimeout(() => {
+        if (DEMO_NODE_CONFIG_TMP.nodeName) {
+          STATUS.node.nodeName = DEMO_NODE_CONFIG_TMP.nodeName
+        }
         DEMO_NODE_CONFIG = { ...DEMO_NODE_CONFIG, ...DEMO_NODE_CONFIG_TMP }
         DEMO_NODE_CONFIG_TMP = {}
         resolve()
